Create db.models.feeds
